stm32f4: Move helper platform headers into separate dir
This helps keep platform interfaces clean
