+++ platform/stm32f4xx/export/aux/pin_descriptor.hpp
