+++ platform/stm32f4xx/export/platform/gpio_device.hpp
@@ -5,7 +5,7 @@
 #define PLATFORM_GPIO_DEVICE_HPP
 
 #include <common/pin.hpp>
-#include "pin_descriptor.hpp"
+#include "aux/pin_descriptor.hpp"
 
 namespace ecl
 {
